requirements.txt
# Core dependencies
streamlit==1.32.0
google-generativeai==0.3.1
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
python-docx==1.0.1

# Vector database and embeddings
huggingface-hub==0.16.4  # This version has cached_download
sentence-transformers==2.2.2
chromadb==0.4.18  # More stable version
torch>=1.6.0
transformers>=4.6.0,<4.30.0

# Cloud storage
google-cloud-storage==2.13.0

# Utilities
tiktoken==0.5.1
json5==0.9.14
